+++ src/CollapsedButton.jsx
@@ -98,7 +98,7 @@ class CollapsedButton extends React.Component {
       return (
         <Button
           {...itemProps}
-          type="outline"
+          type={index === 0 ? 'outline' : 'secondary'}
           size="small"
           disabled={!!item.props.disabled}
         >
@@ -165,7 +165,7 @@ class CollapsedButton extends React.Component {
     if (me.props.type === 'button') {
       return (
         <Dropdown {...dropdownOptions}>
-          <Button className="action" type="outline" size="small">{content}</Button>
+          <Button className="action" type="secondary" size="small">{content}</Button>
         </Dropdown>
       );
     }